+++ CHANGELOG.md
@@ -14,6 +14,20 @@
 
 ### Security
 
+## 0.3.0
+
+### Added
+
+- Add Json Schema for JSON Report Validation
+
+### Changed
+
+- Move from Reporter to [Report API](https://docs.gradle.org/current/javadoc/org/gradle/api/reporting/Report.html)
+
+### Removed
+
+- All Reporter
+
 ## 0.2.0
 
 ### Added